+++ src/EntityGraphQL/Schema/FieldExtensions/ConnectionPaging/ConnectionPagingExtension.cs
@@ -106,7 +106,7 @@ namespace EntityGraphQL.Schema.FieldExtensions
             //      return .... // does the select of only the Conneciton fields asked for
 
             // set up Extension on Edges.Node field to handle the Select() insertion
-            edgesField = returnSchemaType.GetField("edges", null);
+            edgesField = returnSchemaType.GetField(schema.SchemaFieldNamer("Edges"), null);
             // move expression
             edgesField.UpdateExpression(field.Resolve);
             // We steal any previous extensions as they were expected to work on the original Resolve which we moved to Edges