+++ src/EntityGraphQL/Schema/ISchemaProvider.cs
@@ -5,11 +5,11 @@ using EntityGraphQL.Directives;
 namespace EntityGraphQL.Schema
 {
     /// <summary>
-    /// An interface that the EqlCompiler uses to help understand the types it is building against. This abstraction lets us
+    /// An interface that the Compiler uses to help understand the types it is building against. This abstraction lets us
     /// have a simple provider that maps directly to an object as well as other complex providers that read a schema from else where
-    /// and that can map them back to complex expressions. See ObjectSchemaProvider and MappedSchemaProvider for two examples.
+    /// and that can map them back to complex expressions. See SchemaProvider for implementation.
     ///
-    /// It works with type name's as strings because although we ultimately build expressions against actual c# types the provider
+    /// It works with type name's as strings because although we ultimately build expressions against actual C# types the provider
     /// might expose custom names for the underlying type.
     /// </summary>
     public interface ISchemaProvider